OpenWATCOM 1.9 requires this cast.  (I removed it when copying over the wmcbrine code;  apparently,  I was too quick to copy in this case.)

+++ pdcurses/insch.c
@@ -88,7 +88,7 @@ int winsch(WINDOW *win, chtype ch)
     {
         int x2;
 
-        switch (ch)
+        switch ((int)ch)
         {
         case '\t':
             for (x2 = ((x / TABSIZE) + 1) * TABSIZE; x < x2; x++)
